Migrated all getTopic stuff to TopicHelpers. Export 'createTopic' instead, in case people want a topic create utility. Don't export 'getCachedTopic'
