+++ src/components/Subscriber/index.ts
@@ -1,9 +1,2 @@
-export {
-    DefaultMessageType,
-    getTopic,
-    subscribe,
-    TopicSettings,
-    unsubscribe,
-} from './getTopic';
 export { Subscriber, useMsg } from './Subscriber';
 export { useSubscription, UseSubscriptionProps } from './useSubscription';
