+++ src/helpers/TopicHelpers/index.ts
@@ -1,0 +1,3 @@
+export { ManagedTopic } from './ManagedTopic';
+export * from './TopicCache';
+export * from './types';
