+++ src/helpers/TopicHelpers/types.ts
@@ -1,0 +1,12 @@
+import { Message } from 'roslib';
+
+export interface TopicSettings {
+    topic: string;
+    messageType: string;
+    throttleRate?: number;
+    latch?: boolean;
+    queueLength?: number;
+    queueSize?: number;
+}
+
+export type DefaultMessageType = Message;
