+++ src/index.ts
@@ -1,1 +1,2 @@
-export * from "./components";
+export * from './components';
+export { createTopic, DefaultMessageType, TopicSettings } from './helpers/TopicHelpers';
